Move the actual path policies into a public package 'policies'
This will allow them to be used along with the
blueprints.CheckMountpointsPolicy() function outside of images itself.
