+++ pkg/policies/policies_test.go
@@ -1,4 +1,4 @@
-package pathpolicy
+package policies
 
 import "testing"
